Сделал распредление по файлам
